backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
pydantic[email]
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
PyJWT==2.8.0
redis